requirements.txt
matplotlib==3.8.1
networkx==3.2.1
numpy==1.26.2
python-telegram-bot==20.6
python-dotenv==1.0.0